fix(lint): resolve all 10 SwiftLint violations from Phases 2–4
- RunnerModel.swift L111/L124: rename `gh` → `ghStatus` (identifier_name)
- RunnerLifecycleService.swift L15: add doc comment to `shared` (missing_docs)
- RunnerStatusEnricher.swift L16: add doc comment to `APIRunner` (missing_docs)
- RunnerStatusEnricher.swift L40: extract `applyEnrichment` helper to reduce
  cyclomatic complexity from 11 → 6 (cyclomatic_complexity)
- AddRunnerSheet.swift L30: fix colon spacing on `ScopeType` enum case (colon)
- AddRunnerSheet.swift L55: rename `t` → `scopeOption` in ForEach (identifier_name)
- SettingsView.swift L112: remove erroneous space before comma in
  `lifecycleAction` call (comma)
- SettingsView.swift L550: split `RunnerConfigSheet` out into its own file
  `RunnerConfigSheet.swift` to bring SettingsView under the 400-line limit
  (file_length)
